pyproject.toml
[project]
name = "dashlyapi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.22.0",
    "fastapi>=0.119.1",
    "matplotlib>=3.10.7",
    "pandas>=2.3.3",
    "pydantic-settings>=2.11.0",
    "python-multipart>=0.0.20",
    "requests>=2.32.5",
    "uvicorn>=0.38.0",
]
